% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGxEDataset.R
\name{CreateGxEScanDataset}
\alias{CreateGxEScanDataset}
\title{Function to create a dataset for GxEScan}
\usage{
CreateGxEScanDataset(family, covariates, sex = FALSE,
  sexInteraction = FALSE, phenotype = NA)
}
\arguments{
\item{family}{Data frame is the format returned by ReadFamily}

\item{covariates}{Data frame in the format returned by ReadCovariates}

\item{sex}{Indicator to include sex variable from the family file as a covariate}

\item{sexInteraction}{Indicator to use the sex variable from the family file as the interaction covariate.
If sexInteraction is set to TRUE, the value of parameter sex is ignored.}

\item{phenotype}{Name of column in covariate data frame to use as the phenotype. If this value is
set to NA then the phenotype in the family file is used.}
}
\description{
Function to take a family data frame from ReadFamily and a covariate data frame 
from ReadCovariate and create a list containing the matrices needed by the
GxEScan routine
}
