% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEResultTools.R
\name{GrayBackground}
\alias{GrayBackground}
\title{Function to change background of cell to gray}
\usage{
GrayBackground(workbook, sheet, row, column)
}
\arguments{
\item{workbook}{Excel workbook to use}

\item{sheet}{Sheet in workbook to use}

\item{row}{Row number}

\item{column}{Column number}
}
\description{
Function to change background of cell to gray
}
