% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GxEMerge}
\alias{GxEMerge}
\title{Function to merge the results from several GxEScans}
\usage{
GxEMerge(logistic, basefileNames, outfileName)
}
\arguments{
\item{logistic}{Indicator if the outcome was logistic, otherwise outcome was linear}

\item{basefileNames}{List of base file names of results}

\item{outfileName}{Base filename for the merged results}
}
\description{
Function to merge the results from several GxEScans for use with GxEResults
to produce a summary for a genome wide scan
}
