% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEScan.R
\name{GxEScan}
\alias{GxEScan}
\title{Function to perform a GEWIS scan}
\usage{
GxEScan(familyFile, continuous = FALSE, zeroOne = FALSE, pmissing = NA,
  covariateFile, header = TRUE, covsUsed, interactionCov, cmissing = NA,
  useSex = FALSE, sexInteraction = FALSE, phenotype = NA, bedFile,
  mapFile, outfile = "", dg = TRUE, dgxe = TRUE, twodf = TRUE,
  threedf = TRUE, ge = TRUE, caseOnly = TRUE, controlOnly = TRUE,
  dgge = TRUE)
}
\arguments{
\item{familyFile}{Name of file with pedigree data in plink format}

\item{continuous}{Indicator if outcome is continous}

\item{zeroOne}{Indicator if phenotype is coded 0,1. Default is 1,2.}

\item{pmissing}{Value to be treated as missing for phenotype}

\item{covariateFile}{Name of file containing covariate data in plink format}

\item{header}{Indicator if there is a header in the covariate file}

\item{covsUsed}{List of covaraites to be used}

\item{interactionCov}{Covariate to be used in interaction}

\item{cmissing}{Value of covariate to be considered missing}

\item{useSex}{Indicator if value for sex from family file is to be used as a covariate}

\item{sexInteraction}{Indicator if value for sex from family file is to be used as the interacting covariate}

\item{phenotype}{Value from covariate file to be used as phenotype. If the value is NA, the value from
the family file will be used as the phenotype}

\item{bedFile}{Name of measured binary genetic data file}

\item{mapFile}{Name of map file associated with genetic data file}
}
\description{
Function to perform a GEWIS scan
}
