% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEResultTools.R
\name{Process2Step}
\alias{Process2Step}
\title{Function to process a two step test from GxEScan}
\usage{
Process2Step(test1, test2, basefilename, alpha, screen, bins, topnum,
  includeRanks, rankName, ptitle, generatePlot = TRUE, fileOutput = TRUE)
}
\arguments{
\item{test1}{The gxeout output from ProcessGxEOut to be used as the screening test}

\item{test2}{The gxeout output from ProcessGxEOut to be in the second step}

\item{basefilename}{Base filename of output files from GxEScan or GxEMerge also used for plots}

\item{alpha}{Type I error rate}

\item{screen}{Significance level required to be included in second step}

\item{bins}{Number of bins to use in weighted test}

\item{topnum}{Number of top hits to report}

\item{includeRanks}{Indicator to save ranks for Excel output}

\item{rankName}{Name of the column containing the rank}

\item{ptitle}{First part of title for plots}

\item{generatePlot}{Indicator to generate qq and Manhattan plots}

\item{fileOuput}{Indicator to save plots to file - Ignored if generatePlot is FALSE}
}
\description{
Function to take the results for two GxEScan tests that were read in
using the ProcessGxE and produce results for a two step test
}
