% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEResultTools.R
\name{ProcessGxEOut}
\alias{ProcessGxEOut}
\title{Function to process a single gxeout file}
\usage{
ProcessGxEOut(snpinfo, basefilename, extension, df, alpha, topnum, ptitle,
  includeRanks, rankName, generatePlot = TRUE, fileOutput = TRUE)
}
\arguments{
\item{snpinfo}{Data table with SNP information}

\item{basefilename}{Base filename of output files from GxEScan or GxEMerge also used for plots}

\item{extension}{Extenstion to base file name for plot}

\item{df}{Degrees of freedom in test}

\item{alpha}{Type I error rate}

\item{topnum}{Number of top hits to report}

\item{ptitle}{First part of title for plots}

\item{includeRanks}{Indicator to save ranks for Excel output}

\item{rankName}{Name of the column containing the rank}

\item{generatePlot}{Indicator to generate qq and Manhattan plots}

\item{fileOuput}{Indicator to save plots to file - Ignored if generatePlot is FALSE}
}
\description{
Function to process results from GxEScan for a one step test. The
outputs from two tests can be passed to the Process2Step function
to generate results for two step tests.
}
