% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCovariates.R
\name{ReadCovariates}
\alias{ReadCovariates}
\title{Function to read in covariates}
\usage{
ReadCovariates(filename, header = TRUE, used, interaction, cmissing = NA)
}
\arguments{
\item{filename}{Name of covariate file}

\item{header}{Inidicator if the file has a header line. The default value is TRUE.}

\item{used}{List of covariates to keep. When no header is specified this is a list
of column numbers starring with from the third column. When a header
exists, the value is a list of column names. If not value of used is
entered, all covariates in the file will be used.}

\item{interaction}{Column that contains covariate that will be used in iteraction term. This
value will be stored in the file column of the output dataframe. GxEScan
tests the covariate in the last column for gene environment interactions.
When the file has a header this value is the column name. Otherwise it is
the column number. If not interaction is supplied, the first variable in
used list is treated as the interacting covariate.}

\item{cmissing}{Value that is be treated as a missing value. Only one missing value can be
supplied. If no value is entered NA is used.}
}
\value{
A dataframe with the first two columns being family and individual ID, FID and IID.
The remaining column are the value read in and kept as specified by the used
parameter. The last column will be the interacting covariate.
}
\description{
Function to read in covariate information from a file in plink format
}
