% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFam.R
\name{ReadFamily}
\alias{ReadFamily}
\title{Function to read in family data}
\usage{
ReadFamily(filename, continuous = FALSE, zeroOne = FALSE, pmissing = NA)
}
\arguments{
\item{filename}{Name of file to read}

\item{continuous}{Indicator if phenotype is continuous}

\item{zeroOne}{Indicator if dichotomous phenotype is coded 0,1 instead of 1,2}

\item{pmissing}{Value indiating a missing value}
}
\value{
Dataframe with family, individual ID, order records were read in, indicator if
subject is male, and phenotype.
}
\description{
Function to read in a plink formatted family file
}
