% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadGeneticFile}
\alias{ReadGeneticFile}
\title{Function to read a SNP from a measured genetic data file}
\usage{
ReadGeneticFile(BedFilename, numSubjects, MapFilename, SNP)
}
\arguments{
\item{BedFilename}{Name of file with measured genetic data in plink format}

\item{numSubjects}{Number of subjects with data in file}

\item{MapFilename}{Name of map file associated with measured genetic data file}

\item{SNP}{Name of SNP}
}
\value{
Chromosome Number
SNP Name
Location in base pairs
Dosage values
}
\description{
Function to read a SNP from a measured genetic data file
}
