% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScanSNPs}
\alias{ScanSNPs}
\title{Function to perform a GEWIS scan}
\usage{
ScanSNPs(y, x, BedFilename, MapFilename, outFilename, dg, dgxe, twodf, threedf,
  ge, caseOnly, controlOnly, dgge)
}
\arguments{
\item{y}{Vector of outcome values, 0 or 1, all others treated as missing}

\item{x}{Matric of covariates. Last column is tested for interaction with gene}

\item{BedFilename}{Name of file with measured genetic data in plink format}

\item{MapFilename}{Name of map file associated with measured genetic data file}

\item{dg}{Perform D|G test}

\item{dgxe}{Perform Test \deqn{\beta_{x} = 0}}

\item{twodf}{Perform 2df test betaG = 0, betaGxE = 0}

\item{threedf}{Perform 3df test betaG = 0, betaGxE = 0, betaD = 0}

\item{ge}{Perform G|E test}

\item{caseOnly}{Perform G|E test on cases}

\item{controlOnly}{Perform G|E test on cases}

\item{dgge}{Perform 2df test betaG = 0, betaD = 0}
}
\value{
A list containint 5 matrices
Allele frequencies in all subjects, cases, and controls
The number of subjects with complete data
Tne number of cases with complete data
The parameter estimates
The Z statistic for each estimate
The 2df chi-squared statistic for beta_g = 0 and beta_GxE = 0
}
\description{
Function to perform a GEWIS scan
}
