% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEResultTools.R
\name{Write2StepSheet}
\alias{Write2StepSheet}
\title{Function to write 2 step results to Excel worksheet}
\usage{
Write2StepSheet(workbook, testName, testType1, testType2, basefilename,
  extension, description, m, sig, topHits, includeRanks, testRanks, rankOrder)
}
\arguments{
\item{workbook}{Workbook to write results to}

\item{testName}{Name of column in testRanks for the test being output}

\item{testType1}{Type of statistic for step 1, e.g., ttest, chisq}

\item{testType2}{Type of statistic for step 2, e.g., ttest, chisq}

\item{basefilename}{Base filename of plots}

\item{extension}{Extension to base filename for plot
Also used to create the name of the sheet created for the results}

\item{description}{Description of test}

\item{m}{Number of SNPs that test passed step 1}

\item{sig}{Required significance level}

\item{topHits}{Data table of top hits for test}

\item{includeRanks}{Indicator to include ranks on sheet}

\item{testRanks}{Ranks for all tests performed}

\item{rankOrder}{Order to display rank columns}
}
\description{
Function to write 2 step results to Excel worksheet
}
