% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEResultTools.R
\name{WriteExhaustiveSheet}
\alias{WriteExhaustiveSheet}
\title{Function to write exhaustive test results to Excel spreedsheet}
\usage{
WriteExhaustiveSheet(workbook, testName, basefilename, extension, description,
  M, sig, topHits, includeRanks, numMethods, testRanks, rankOrder)
}
\arguments{
\item{workbook}{Workbook to write results to}

\item{testName}{Name of column in testRanks for the test being output
Also is the name of the sheet created for the results}

\item{basefilename}{Base filename of plots}

\item{extension}{Extension to base filename for plot}

\item{description}{Description of test}

\item{M}{Number of SNPs that test converged for}

\item{sig}{Required significance level}

\item{topHits}{Data table of top hits for test}

\item{includeRanks}{Indicator to include ranks on sheet}

\item{numMethods}{Number of methods included in analysis}

\item{testRanks}{Ranks for all tests performed}

\item{rankOrder}{Order to display rank columns}
}
\description{
Function to write exhaustive test results to Excel spreedsheet
}
