% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEResultTools.R
\name{WriteSummaryWorksheet}
\alias{WriteSummaryWorksheet}
\title{Function to write GxEScan summary sheet}
\usage{
WriteSummaryWorksheet(workbook, exhaustiveTests, exhaustiveNames, numSNPs1,
  sig1, twoStepTests, twoStepNames, numSNPs2, sig2, bins, step1Test,
  stacked = TRUE, descrip2side = TRUE)
}
\arguments{
\item{exhaustiveTests}{List of indicator of exhaustive tests performed}

\item{exhaustiveNames}{Names of exhaustive test names}

\item{numSNPs1}{Number of SNPs that converged for each test}

\item{sig1}{Overall significance level for each test}

\item{twoStepTests}{Indicator of two step tests performed}

\item{twoStepNames}{Names of two step tests}

\item{numSNPs2}{Number of SNPs that made it to step 2}

\item{sig2}{Overall significance level for two step test}

\item{bins}{Number of bins for weighted tests}

\item{step1Test}{Tests performed for step 1 of two step test}

\item{stacked}{Indicator if two step steps are underneath exhaustive test, otherwise side by side}

\item{filename}{Excel workbook to write summary sheet to}

\item{decrip2side}{Inidicator if method decriptions are written to side of table, otherwise written below}
}
\description{
Function to write GxEScan summary sheet
listing tests performed, number of SNPs tests converged for,
overall significance level, and required significance
}
