% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEResultTools.R
\name{WriteWeightedSheet}
\alias{WriteWeightedSheet}
\title{Function to write weighted 2 step results to Excel worksheet}
\usage{
WriteWeightedSheet(workbook, basefilename, extension, description, topHits,
  testType1, testType2)
}
\arguments{
\item{workbook}{Workbook to write results to}

\item{basefilename}{Base filename of plots}

\item{extension}{Extension to base filename for plot
Also used to create the name of the sheet created for the results}

\item{description}{Description of test}

\item{topHits}{Data table of top hits for test}

\item{testType1}{Type of statistic for step 1, e.g., ttest, chisq}

\item{testType2}{Type of statistic for step 2, e.g., ttest, chisq}
}
\description{
Function to write weighted 2 step results to Excel worksheet
}
