% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEResults.R
\name{GxEResults}
\alias{GxEResults}
\title{Function to process results from GxEScan}
\usage{
GxEResults(homedir, basefilename, alpha = 0.05, exhaustiveTests,
  topnum = 10, includeRanks = FALSE, bins, screens, twoStepTests)
}
\arguments{
\item{basefilename}{Base filename of output files from GxEScan or GxEMerge}

\item{alpha}{Type I error rate}

\item{topnum}{Number of top hits to report}

\item{includeRanks}{Include ranks}
}
\description{
Function to process results from GxEScan to perform the two-step steps
and produce summaries and plots in an Excel spreadsheet
}
